[FIX] Move vals correction to virtual relations.
